tools/intel_i2c_slave: Add a bitbanged i2c slave
Add a crude bitbanged i2c slave that can be used for testing/debugging
eg. GMBUS. Just hook up a VGA/HDMI cable between two machines,
run the i2c slave on one and a i2c master on the other.

For now this slave more or less pretends to be an 8 byte long
EEPROM at the EDID slave address. Should probably make it simulate
a full EDID with segment support and all...

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>