tools/intel_gtt_dump: Add a tool for dumping gtt contents as pngs
When looking into plane scanout issues it can be helpful to dump the
memory supposedly being scanned out into an image. It can then be
compared with what the display is showing. Add a small tool to do that.

Note that with a PAT enabled CPU the kernel is being a douche and
prevents access to normal RAM via /dev/mem. That check needs to be
patched out from the kernel before the tool actually works.

TODO: Add a mode that reads out the plane registers and figures
out the dimensions/format/etc. automagically

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>