tools/intel_gtt_dump: Don't accept offset 0
When run withouth any arguments we default to offset 0. Often that is
point at stolen memory and trying to access that can hard hang the box.
Require that the user specify at least some non-zero offset before
we go poking at the memory. Ideally we'd check that we're dealing with
normal system RAM, but for a developer debug tool that seems too much
work for little gain.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>